+++ lib/resources/api_keys.php
@@ -4,6 +4,7 @@ class APIKeys
 {
   protected
     $name,
+    $scopes,
     $base_endpoint,
     $endpoint,
     $client;
@@ -33,17 +34,33 @@ class APIKeys
   }
   
   public function get(){
+    $this->endpoint = $this->base_endpoint;
     return $this->client->getRequest($this);
   }
   
-  public function post($name){
+  public function post($name, $scopes = NULL){
+    $this->endpoint = $this->base_endpoint;
     $data = array(
       'name' => $name,
     );
+    if ($scopes)
+        $data["scopes"] = $scopes;
+    
     return $this->client->postRequest($this, $data);
   }
   
+  public function put($api_key_id, $name, $scopes){
+    $this->endpoint = $this->base_endpoint;
+    $data = array(
+      'name' => $name,
+      'scopes' => $scopes
+    );
+    $this->endpoint = $this->base_endpoint . "/" . $api_key_id;
+    return $this->client->putRequest($this, $data);
+  }
+  
   public function patch($api_key_id, $name){
+    $this->endpoint = $this->base_endpoint;
     $data = array(
       'name' => $name,
     );
@@ -52,6 +69,7 @@ class APIKeys
   }  
   
   public function delete($api_key_id){
+    $this->endpoint = $this->base_endpoint;
     $this->endpoint = $this->base_endpoint . "/" . $api_key_id;
     return $this->client->deleteRequest($this);
   }