+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [v4.0.4] - (2016-02-18) ##
+### Added
+- Ability to add scopes to API Keys endpoint [POST]
+
+## [v4.0.3] - (2016-02-18) ##
+### Added
+- API Keys endpoint [PUT]
+
+## [v4.0.2] - (2015-12-15) ##
+### Added
+- Tests for API Keys endpoint [POST, PATCH, DELETE]
+
 ## [v4.0.1] - (2015-12-03) ##
 ### Fixed
 - HTTP 406 Not Acceptable Errors [#177](https://github.com/sendgrid/sendgrid-php/issues/177)